Changes to custom Reports to support outputting to Excel/Word etc.
This is the first part of the fix in which we're moving the old vb6 activex client upto the server and convert over to .NET (previous check-in). I've then had to substantially rework how the custom reports run . 

A small history lesson. WE used to build up strings put them into input controls stream them to the client and then there was some javascript that would read these input in and put into a table (or an activex grid if you go further back). These values would also get read into a hidden grid which was then traversed to take the values and put into the client activex ready for output. This was very cumbersome.

In order to get this all running on the server I've modified things to we now populate a datatable and bind this directly to a gridview. This datatable is now also passed to the download_ouput in the homecontroller where the data is sent to the converted cleint side code.

At present this code still relies on having Office installed on the server, but there will be another imminent check-in where these objects are replaced so don't woryy too much at this stage about testing the actual ouput.

I've also changed some of the multi-dimensional arrays to be strongly typed collections as understanding that part of the code was an absolute mare.

2 known bugs which I'm going to raise separatly 1) the group with next does not put linespaces in and secondly the columns need some kind of formatting to the right size.

+++ Components/COAInt_Server/clsSettings.vb
@@ -16,34 +16,34 @@ Public Class clsSettings
   End Function
 
 
-  Public Function GetWordColourIndex(ByRef lngColourValue As Integer) As Integer
+	Public Function GetWordColourIndex(ByRef lngColourValue As Long) As Integer
 
-    Dim rsTemp As ADODB.Recordset
-    Dim strSQL As String
+		Dim rsTemp As ADODB.Recordset
+		Dim strSQL As String
 
-    On Error GoTo LocalErr
+		On Error GoTo LocalErr
 
-    strSQL = "SELECT WordColourIndex FROM ASRSysColours " & " WHERE ColValue = " & CStr(lngColourValue)
-    rsTemp = datGeneral.GetReadOnlyRecords(strSQL)
+		strSQL = "SELECT WordColourIndex FROM ASRSysColours " & " WHERE ColValue = " & CStr(lngColourValue)
+		rsTemp = datGeneral.GetReadOnlyRecords(strSQL)
 
-    With rsTemp
-      If Not .BOF And Not .EOF Then
-        GetWordColourIndex = rsTemp.Fields("WordColourIndex").Value
-      Else
-        GetWordColourIndex = 0
-      End If
-    End With
+		With rsTemp
+			If Not .BOF And Not .EOF Then
+				GetWordColourIndex = rsTemp.Fields("WordColourIndex").Value
+			Else
+				GetWordColourIndex = 0
+			End If
+		End With
 
-    rsTemp.Close()
-    'UPGRADE_NOTE: Object rsTemp may not be destroyed until it is garbage collected. Click for more: 'ms-help://MS.VSCC.v90/dv_commoner/local/redirect.htm?keyword="6E35BFF6-CD74-4B09-9689-3E1A43DF8969"'
-    rsTemp = Nothing
+		rsTemp.Close()
+		'UPGRADE_NOTE: Object rsTemp may not be destroyed until it is garbage collected. Click for more: 'ms-help://MS.VSCC.v90/dv_commoner/local/redirect.htm?keyword="6E35BFF6-CD74-4B09-9689-3E1A43DF8969"'
+		rsTemp = Nothing
 
-    Exit Function
+		Exit Function
 
 LocalErr:
-    GetWordColourIndex = 0
+		GetWordColourIndex = 0
 
-  End Function
+	End Function
 
 
   Public Function GetEmailGroupAddresses(ByRef lngGroupID As Integer) As String
